[Refactor Google APIs into separate modules]
Detailed Changes:
- Moved the Google Calendar-related code from `google.py` to a new `gcal.py` module
- Moved the Gmail-related code from `google.py` to a new `gmail.py` module
- Moved the shared Google API authentication code to a new `google_shared.py` module
- Updated the imports and references throughout the codebase to use the new module structure

Impact: This change improves the code organization and maintainability by separating the Google API-related functionality into more focused modules. It also centralizes the shared authentication logic, making it easier to manage and update in the future.

Additional Notes: The overall functionality of the tools remains the same, but the internal structure has been reorganized to follow a more modular approach.
